refactor(server): move challenge interface definitions to types.ts
