+++ server/challenges/types.ts
